refactor: parseSnsSwapSaleBuyerCount
